added fixes all over the place

+++ duckdb-be/user_data.py
@@ -32,6 +32,7 @@ create_tables(cursor, 'data_scripts/postgres/create_tables.sql')
 
 # Reading data
 user_df = pd.read_csv("data/raw/user.csv")
+
 address_df = pd.read_csv("data/raw/address.csv")
 account_df = pd.read_csv("data/raw/account.csv")
 
@@ -39,7 +40,7 @@ account_df = pd.read_csv("data/raw/account.csv")
 address_df["per_capita_income"] = address_df["per_capita_income"].replace('[\$,]', '', regex=True).astype(float)
 account_df["yearly_income"] = account_df["yearly_income"].replace('[\$,]', '', regex=True).astype(float)
 account_df["total_debt"] = account_df["total_debt"].replace('[\$,]', '', regex=True).astype(float)
-
+account_df["credit_score"] = account_df["credit_score"].astype(float)
 insert_data(user_df, "User")      
 insert_data(address_df, "Address")
 insert_data(account_df, "Account")
